@@ -13,18 +13,22 @@
 		}
 	},
 	"dependencies": {
-		"@angular/animations": "~5.0.0",
-		"@angular/common": "~5.0.0",
-		"@angular/compiler": "~5.0.0",
-		"@angular/core": "~5.0.0",
-		"@angular/forms": "~5.0.0",
-		"@angular/http": "~5.0.0",
-		"@angular/platform-browser": "~5.0.0",
-		"@angular/platform-browser-dynamic": "~5.0.0",
-		"@angular/router": "~5.0.0",
+		"@angular/animations": "~5.1.0",
+		"@angular/common": "~5.1.0",
+		"@angular/compiler": "~5.1.0",
+		"@angular/core": "~5.1.0",
+		"@angular/forms": "~5.1.0",
+		"@angular/http": "~5.1.0",
+		"@angular/platform-browser": "~5.1.0",
+		"@angular/platform-browser-dynamic": "~5.1.0",
+		"@angular/router": "~5.1.0",
+		"@ngrx/store": "^4.1.1",
+		"@ngrx/store-devtools": "^4.1.1",
 		"allow-publish": "^1.0.4",
-		"nativescript-angular": "~5.0.0",
+		"nativescript-angular": "^5.1.0",
+		"nativescript-checkbox": "^3.0.3",
 		"nativescript-theme-core": "~1.0.4",
+		"ngrx-store-freeze": "^0.2.0",
 		"reflect-metadata": "~0.1.8",
 		"rxjs": "~5.5.2",
 		"tns-core-modules": "~3.4.0",
