@@ -1,7 +1,7 @@
 // this import should be first in order to load some required settings (like globals and reflect-metadata)
 import { platformNativeScriptDynamic } from 'nativescript-angular/platform';
 
-import { AppNativeModule } from './app/app.module';
+import { AppNativeModule } from './app/app.native.module';
 
 // A traditional NativeScript application starts by initializing global objects,
 //  setting up global CSS rules, creating, and navigating to the main page.
