@@ -1,9 +1,10 @@
 import { NgModule } from '@angular/core';
-import { NativeScriptRouterModule } from 'nativescript-angular/router';
 import { Routes } from '@angular/router';
+import { NativeScriptRouterModule } from 'nativescript-angular/router';
 
 const routes: Routes = [
-
+  { path: '', loadChildren: './app/home/home.module#HomeModule' },
+  { path: '**', redirectTo: '' },
 ];
 
 @NgModule({
