@@ -1,24 +1,19 @@
 import { NgModule, NO_ERRORS_SCHEMA } from '@angular/core';
 import { NativeScriptModule } from 'nativescript-angular/nativescript.module';
+import { TNSCheckBoxModule } from 'nativescript-checkbox/angular';
+
+import { AppModule } from './app.module';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-
-// Uncomment and add to NgModule imports if you need to use two-way binding
-// import { NativeScriptFormsModule } from "nativescript-angular/forms";
-
-// Uncomment and add to NgModule imports  if you need to use the HTTP wrapper
-// import { NativeScriptHttpModule } from "nativescript-angular/http";
 
 @NgModule({
   bootstrap: [
     AppComponent
   ],
   imports: [
+    AppModule,
     NativeScriptModule,
-    AppRoutingModule
-  ],
-  declarations: [
-    AppComponent
+    TNSCheckBoxModule
   ],
   schemas: [
     NO_ERRORS_SCHEMA
