@@ -1,13 +1,16 @@
 import { NgModule } from '@angular/core';
 import { ServerModule } from '@angular/platform-server';
 import { ModuleMapLoaderModule } from '@nguniversal/module-map-ngfactory-loader';
+import { ServiceWorkerModule } from '@angular/service-worker';
+import { environment } from '../environments/environment';
 
-import { AppModule } from '@app/app.module';
-import { AppComponent } from '@app/app.component';
+import { AppModule } from './app.module';
+import { AppComponent } from './app.component';
 
 @NgModule({
   imports: [
     AppModule,
+    ServiceWorkerModule.register('/ngsw-worker.js', { enabled: environment.production }),
     ServerModule,
     ModuleMapLoaderModule
   ],
