@@ -1,6 +1,19 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { NO_ERRORS_SCHEMA, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { Observable } from 'rxjs/Observable';
+import { Store } from '@ngrx/store';
+import 'rxjs/add/observable/of';
 
 import { TodoListComponent } from './todo-list.component';
+
+const mockStore = {
+  select: jasmine.createSpy('select').and.returnValues(
+    Observable.of([{ id: 0, text: 'foo', completed: false }]),
+    Observable.of('')
+  ),
+  dispatch: jasmine.createSpy('dispatch')
+};
 
 describe('TodoListComponent', () => {
   let component: TodoListComponent;
@@ -8,6 +21,14 @@ describe('TodoListComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
+      imports: [ FormsModule ],
+      providers: [
+        { provide: Store, useValue: mockStore }
+      ],
+      schemas: [
+        NO_ERRORS_SCHEMA,
+        CUSTOM_ELEMENTS_SCHEMA
+      ],
       declarations: [ TodoListComponent ]
     })
     .compileComponents();
@@ -19,7 +40,7 @@ describe('TodoListComponent', () => {
     fixture.detectChanges();
   });
 
-  it('should create', () => {
+  it('should create component', () => {
     expect(component).toBeTruthy();
   });
 });
