@@ -1,16 +1,18 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild, ElementRef, ChangeDetectionStrategy } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import { take } from 'rxjs/operators';
 import { Store } from '@ngrx/store';
 
-import { Todo, TodosState } from '@app/todos/shared';
-import { TodosActionTypes, getTodoListState, getTodosFormInputState } from '@app/todos/store';
+import { Todo, TodosState } from '../shared';
+import { TodosActionTypes, getTodoListState, getTodosFormInputState } from '../store';
+import { filter } from 'rxjs/operators';
 
 @Component({
   moduleId: module.id,
   selector: 'app-todo-list',
   templateUrl: 'todo-list.component.html',
-  styleUrls: [ 'todo-list.component.scss' ]
+  styleUrls: [ 'todo-list.component.scss' ],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class TodoListComponent implements OnInit {
   /**
